--- packages/hundredandten-engine/src/hundredandten/engine/trumps.py
@@ -1,5 +1,0 @@
-"""Re-export trump helpers from hundredandten-deck."""
-
-from hundredandten.deck import bleeds, trumps
-
-__all__ = ["bleeds", "trumps"]
